# backend/static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Deliver00 API",
      "description": "API for managing users and cargo in a delivery system",
      "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/api/v1",
    "schemes": ["http"],
    "paths": {
      "/auth/login": {
        "post": {
          "summary": "Login a user",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          ],
          "responses": {
            "200": { "description": "Login successful" },
            "401": { "description": "Invalid credentials" }
          }
        }
      },
      "/auth/signup": {
        "post": {
          "summary": "Register a new user",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "role": { "type": "string" }
                }
              }
            }
          ],
          "responses": {
            "201": { "description": "User created successfully" },
            "400": { "description": "Email already registered" }
          }
        }
      },
      "/users": {
        "get": {
          "summary": "Get all users",
          "responses": {
            "200": { "description": "List of users" }
          }
        }
      },
      "/parcels": {
        "post": {
          "summary": "Create a new parcel",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "origin_pin": { "type": "string" },
                  "destination_pin": { "type": "string" },
                  "weight_kg": { "type": "number" },
                  "description": { "type": "string" },
                  "user_id": { "type": "string" }
                }
              }
            }
          ],
          "responses": {
            "201": { "description": "Parcel created successfully" },
            "400": { "description": "Invalid input" }
          }
        }
      },
      "/users/{user_id}/parcels": {
        "get": {
          "summary": "Get all parcels for a user",
          "parameters": [
            {
              "name": "user_id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": { "description": "List of user's parcels" },
            "404": { "description": "User not found" }
          }
        }
      },
      "/parcels/{parcel_id}": {
        "get": {
          "summary": "Get parcel by ID",
          "parameters": [
            {
              "name": "parcel_id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": { "description": "Parcel details" },
            "404": { "description": "Parcel not found" }
          }
        }
      },
      "/parcels/{parcel_id}/cancel": {
        "put": {
          "summary": "Cancel a parcel",
          "parameters": [
            {
              "name": "parcel_id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": { "description": "Parcel canceled" },
            "404": { "description": "Parcel not found" }
          }
        }
      },
      "/parcel/{parcel_id}": {
      "put": {
        "summary": "Update parcel details",
        "parameters": [
          {
            "name": "parcel_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "origin_pin": { "type": "string" },
                "destination_pin": { "type": "string" },
                "weight_kg": { "type": "number" },
                "description": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Parcel details updated successfully" },
          "404": { "description": "Parcel not found" }
        }
      }
    },
      "/parcels/{parcel_id}/status": {
        "put": {
          "summary": "Update parcel status (Admin only)",
          "parameters": [
            {
              "name": "parcel_id",
              "in": "path",
              "required": true,
              "type": "integer"
            },
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string" },
                  "role": { "type": "string" }
                }
              }
            }
          ],
          "responses": {
            "200": { "description": "Parcel status updated" },
            "403": { "description": "Admin privileges required" },
            "404": { "description": "Parcel not found" }
          }
        }
      }
    }
  }
  